Fixed the issue of not combile when test case depend on other class
Took 45 minutes
